refactor Transformer
