+++ server/src/main/java/org/apache/iotdb/db/query/udf/core/transformer/binary/ArithmeticAdditionTransformer.java
@@ -17,25 +17,19 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.query.udf.core.transformer;
+package org.apache.iotdb.db.query.udf.core.transformer.binary;
 
 import org.apache.iotdb.db.query.udf.core.reader.LayerPointReader;
-import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 
-public abstract class CompareBinaryTransformer extends BinaryTransformer {
+public class ArithmeticAdditionTransformer extends ArithmeticBinaryTransformer {
 
-  protected CompareBinaryTransformer(
+  public ArithmeticAdditionTransformer(
       LayerPointReader leftPointReader, LayerPointReader rightPointReader) {
     super(leftPointReader, rightPointReader);
   }
 
   @Override
-  protected TransformerType getTransformerType() {
-    return TransformerType.Comparative;
-  }
-
-  @Override
-  public TSDataType getDataType() {
-    return TSDataType.BOOLEAN;
+  protected double evaluateDouble(double leftOperand, double rightOperand) {
+    return leftOperand + rightOperand;
   }
 }
