+++ server/src/main/java/org/apache/iotdb/db/query/udf/core/transformer/binary/ArithmeticDivisionTransformer.java
@@ -17,24 +17,19 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.query.udf.core.transformer;
+package org.apache.iotdb.db.query.udf.core.transformer.binary;
 
 import org.apache.iotdb.db.query.udf.core.reader.LayerPointReader;
 
-public class CompareEqualToTransformer extends CompareBinaryTransformer {
+public class ArithmeticDivisionTransformer extends ArithmeticBinaryTransformer {
 
-  public CompareEqualToTransformer(
+  public ArithmeticDivisionTransformer(
       LayerPointReader leftPointReader, LayerPointReader rightPointReader) {
     super(leftPointReader, rightPointReader);
   }
 
   @Override
-  protected TransformerType getTransformerType() {
-    return TransformerType.EqNeq;
-  }
-
-  @Override
-  protected boolean evaluateBoolean(double leftOperand, double rightOperand) {
-    return Double.compare(leftOperand, rightOperand) == 0;
+  protected double evaluateDouble(double leftOperand, double rightOperand) {
+    return leftOperand / rightOperand;
   }
 }
