+++ server/src/main/java/org/apache/iotdb/db/query/udf/core/transformer/binary/ArithmeticModuloTransformer.java
@@ -17,19 +17,19 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.query.udf.core.transformer;
+package org.apache.iotdb.db.query.udf.core.transformer.binary;
 
 import org.apache.iotdb.db.query.udf.core.reader.LayerPointReader;
 
-public class ArithmeticDivisionTransformer extends ArithmeticBinaryTransformer {
+public class ArithmeticModuloTransformer extends ArithmeticBinaryTransformer {
 
-  public ArithmeticDivisionTransformer(
+  public ArithmeticModuloTransformer(
       LayerPointReader leftPointReader, LayerPointReader rightPointReader) {
     super(leftPointReader, rightPointReader);
   }
 
   @Override
   protected double evaluateDouble(double leftOperand, double rightOperand) {
-    return leftOperand / rightOperand;
+    return leftOperand % rightOperand;
   }
 }
