+++ server/src/main/java/org/apache/iotdb/db/query/udf/core/transformer/binary/ArithmeticMultiplicationTransformer.java
@@ -17,18 +17,19 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.query.udf.core.transformer;
+package org.apache.iotdb.db.query.udf.core.transformer.binary;
 
 import org.apache.iotdb.db.query.udf.core.reader.LayerPointReader;
 
-public class LogicOrTransformer extends LogicBinaryTransformer {
+public class ArithmeticMultiplicationTransformer extends ArithmeticBinaryTransformer {
 
-  public LogicOrTransformer(LayerPointReader leftPointReader, LayerPointReader rightPointReader) {
+  public ArithmeticMultiplicationTransformer(
+      LayerPointReader leftPointReader, LayerPointReader rightPointReader) {
     super(leftPointReader, rightPointReader);
   }
 
   @Override
-  protected boolean evaluateBoolean(boolean leftOperand, boolean rightOperand) {
-    return leftOperand || rightOperand;
+  protected double evaluateDouble(double leftOperand, double rightOperand) {
+    return leftOperand * rightOperand;
   }
 }
