+++ server/src/main/java/org/apache/iotdb/db/query/udf/core/transformer/binary/CompareLessEqualTransformer.java
@@ -17,18 +17,19 @@
  * under the License.
  */
 
-package org.apache.iotdb.db.query.udf.core.transformer;
+package org.apache.iotdb.db.query.udf.core.transformer.binary;
 
 import org.apache.iotdb.db.query.udf.core.reader.LayerPointReader;
 
-public class LogicAndTransformer extends LogicBinaryTransformer {
+public class CompareLessEqualTransformer extends CompareBinaryTransformer {
 
-  public LogicAndTransformer(LayerPointReader leftPointReader, LayerPointReader rightPointReader) {
+  public CompareLessEqualTransformer(
+      LayerPointReader leftPointReader, LayerPointReader rightPointReader) {
     super(leftPointReader, rightPointReader);
   }
 
   @Override
-  protected boolean evaluateBoolean(boolean leftOperand, boolean rightOperand) {
-    return leftOperand && rightOperand;
+  protected boolean evaluateBoolean(double leftOperand, double rightOperand) {
+    return Double.compare(leftOperand, rightOperand) <= 0;
   }
 }
